20230601
20230601

+++ wp-content/plugins/crud/console/migrations/m230318_162338_create_wp_order_addcolumn.php
@@ -29,7 +29,10 @@ class m230318_162338_create_wp_order_addcolumn extends Migration
     // Use up()/down() to run migration code without a transaction.
     public function up()
     {
-        $this->dropTable('{{%order}}');
+        if($this->tableExists('{{%order}}')){
+            $this->dropTable('{{%order}}');
+        }
+
         $tableOptions = null;
         if ($this->db->driverName === 'mysql') {
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_general_ci ENGINE=InnoDB COMMENT="订单表"';
@@ -57,9 +60,16 @@ class m230318_162338_create_wp_order_addcolumn extends Migration
 
     public function down()
     {
+        $this->dropTable('{{%order}}');
         echo "m230318_162338_create_wp_order_addcolumn cannot be reverted.\n";
 
         return false;
     }
+
+    public function tableExists($tableName){
+        $tableSchema = \Yii::$app->db->schema->getTableSchema($tableName);
+        return ($tableSchema !== null);
+    }
+
 
 }
